+++ variants/tests/test_forms.py
@@ -23,6 +23,10 @@ class TestFormBase(TestCase):
                 hpo_id="HP:0000003",
                 name="Disease 1;;Alternative Description",
             ),
+            HpoFactory(database_id="DECIPHER:1", hpo_id="HP:0000003", name="Disease 2",),
+            HpoFactory(database_id="DECIPHER:1", hpo_id="HP:0000004", name="Disease 2",),
+            HpoFactory(database_id="ORPHA:1", hpo_id="HP:0000004", name="Disease 3",),
+            HpoFactory(database_id="ORPHA:1", hpo_id="HP:0000005", name="Disease 3",),
         ]
         self.maxDiff = None
 
@@ -53,20 +57,109 @@ class TestFilterForm(TestFormBase):
         self.assertTrue(form.is_valid())
         self.assertEqual(form.cleaned_data["genomic_region"], [("X", None, None)])
 
-    def test_omim_to_hpo_conversion(self):
+    def test_hpo_terms_plain(self):
         form_data = vars(
             FormDataFactory(
                 names=self.variant_set.case.get_members(),
-                prio_hpo_terms="; ".join([self.hponame.hpo_id, self.hpos[0].database_id]),
+                prio_hpo_terms="; ".join([self.hponame.hpo_id]),
             )
         )
         form = FilterForm(form_data, case=self.variant_set.case, user=0)
         self.assertTrue(form.is_valid())
-        self.assertEqual(
+        self.assertListEqual(
+            form.cleaned_data["prio_hpo_terms"], sorted([self.hponame.hpo_id]),
+        )
+        self.assertListEqual(
+            form.cleaned_data["prio_hpo_terms_curated"], sorted([self.hponame.hpo_id]),
+        )
+
+    def test_hpo_terms_omim_to_hpo_conversion(self):
+        form_data = vars(
+            FormDataFactory(
+                names=self.variant_set.case.get_members(),
+                prio_hpo_terms="; ".join([self.hpos[0].database_id]),
+            )
+        )
+        form = FilterForm(form_data, case=self.variant_set.case, user=0)
+        self.assertTrue(form.is_valid())
+        self.assertListEqual(
+            form.cleaned_data["prio_hpo_terms"], sorted([self.hpos[0].database_id]),
+        )
+        self.assertListEqual(
+            form.cleaned_data["prio_hpo_terms_curated"],
+            sorted([self.hpos[0].hpo_id, self.hpos[1].hpo_id]),
+        )
+
+    def test_hpo_terms_decipher_to_hpo_conversion(self):
+        form_data = vars(
+            FormDataFactory(
+                names=self.variant_set.case.get_members(),
+                prio_hpo_terms="; ".join([self.hpos[2].database_id]),
+            )
+        )
+        form = FilterForm(form_data, case=self.variant_set.case, user=0)
+        self.assertTrue(form.is_valid())
+        self.assertListEqual(
+            form.cleaned_data["prio_hpo_terms"], sorted([self.hpos[2].database_id]),
+        )
+        self.assertListEqual(
+            form.cleaned_data["prio_hpo_terms_curated"],
+            sorted([self.hpos[2].hpo_id, self.hpos[3].hpo_id]),
+        )
+
+    def test_hpo_terms_orpha_to_hpo_conversion(self):
+        form_data = vars(
+            FormDataFactory(
+                names=self.variant_set.case.get_members(),
+                prio_hpo_terms="; ".join([self.hpos[4].database_id]),
+            )
+        )
+        form = FilterForm(form_data, case=self.variant_set.case, user=0)
+        self.assertTrue(form.is_valid())
+        self.assertListEqual(
+            form.cleaned_data["prio_hpo_terms"], sorted([self.hpos[4].database_id]),
+        )
+        self.assertListEqual(
+            form.cleaned_data["prio_hpo_terms_curated"],
+            sorted([self.hpos[4].hpo_id, self.hpos[5].hpo_id]),
+        )
+
+    def test_hpo_terms_mixed(self):
+        form_data = vars(
+            FormDataFactory(
+                names=self.variant_set.case.get_members(),
+                prio_hpo_terms="; ".join(
+                    [
+                        self.hponame.hpo_id,
+                        self.hpos[0].database_id,
+                        self.hpos[2].database_id,
+                        self.hpos[4].database_id,
+                    ]
+                ),
+            )
+        )
+        form = FilterForm(form_data, case=self.variant_set.case, user=0)
+        self.assertTrue(form.is_valid())
+        self.assertListEqual(
             form.cleaned_data["prio_hpo_terms"],
-            sorted([self.hponame.hpo_id, self.hpos[0].database_id]),
+            sorted(
+                [
+                    self.hponame.hpo_id,
+                    self.hpos[0].database_id,
+                    self.hpos[2].database_id,
+                    self.hpos[4].database_id,
+                ]
+            ),
         )
         self.assertEqual(
             form.cleaned_data["prio_hpo_terms_curated"],
-            sorted([self.hponame.hpo_id, self.hpos[0].hpo_id, self.hpos[1].hpo_id]),
+            sorted(
+                [
+                    self.hponame.hpo_id,
+                    self.hpos[0].hpo_id,
+                    self.hpos[1].hpo_id,
+                    self.hpos[3].hpo_id,
+                    self.hpos[5].hpo_id,
+                ]
+            ),
         )