+++ variants/views.py
@@ -2104,19 +2104,18 @@ class CaseLoadPrefetchedFilterView(
             if hpo.startswith("HP"):
                 matches = HpoName.objects.filter(hpo_id=hpo)
                 hpoterms[hpo] = matches.first().name if matches else "unknown HPO term"
-            elif hpo.startswith("OMIM"):
+            else:
                 matches = (
                     Hpo.objects.filter(database_id=hpo)
                     .values("database_id")
                     .annotate(names=ArrayAgg("name"))
                 )
-                hpoterms[hpo] = (
-                    re.sub(r"^[#%]?\d{6} ", "", matches.first()["names"][0]).split(";;")[0]
-                    if matches
-                    else "unknown OMIM term"
-                )
-            else:
-                hpoterms[hpo] = "unknown term"
+                if matches:
+                    hpoterms[hpo] = re.sub(r"^[#%]?\d+ ", "", matches.first()["names"][0]).split(
+                        ";;"
+                    )[0]
+                else:
+                    hpoterms[hpo] = "unknown term"
 
         user = self.request.user
         if settings.KIOSK_MODE:
@@ -3845,29 +3844,31 @@ class HpoTermsApiView(LoginRequiredMixin, View):
         if not query:
             return HttpResponse({}, content_type="application/json")
         hpo = HpoName.objects.filter(Q(hpo_id__icontains=query) | Q(name__icontains=query))[:10]
-        omim = (
-            Hpo.objects.filter(
-                (Q(database_id__startswith="OMIM") & Q(database_id__icontains=query))
-                | Q(name__icontains=query)
-            )
+        omim_decipher_orpha = (
+            Hpo.objects.filter(Q(database_id__icontains=query) | Q(name__icontains=query))
             .values("database_id")
             .distinct()[:10]
         )
         result = []
         for h in hpo:
             result.append({"id": h.hpo_id, "name": h.name})
-        for o in omim:
+        for o in omim_decipher_orpha:
             names = []
-            # Query database again to get all possible names for an OMIM id
+            # Query database again to get all possible names for an OMIM/DECIPHER/ORPHA id
             for name in (
                 Hpo.objects.filter(database_id=o["database_id"])
                 .values("database_id")
                 .annotate(names=ArrayAgg("name"))[0]["names"]
             ):
-                for n in re.sub(r"^[#%]?\d{6} ", "", name).split(";;"):
-                    if n not in names:
-                        names.append(n)
+                if o["database_id"].startswith("OMIM"):
+                    for n in re.sub(r"^[#%]?\d{6} ", "", name).split(";;"):
+                        if n not in names:
+                            names.append(n)
+                else:
+                    if name not in names:
+                        names.append(name)
             result.append({"id": o["database_id"], "name": ";;".join(names)})
+
         return HttpResponse(json.dumps(result), content_type="application/json")