+++ HISTORY.rst
@@ -30,6 +30,7 @@ End-User Summary
 - Displaying warning if priorization is not enabled when entering HPO terms.
 - Added possibility to import "extra annotations" for display along with the variants.
 - On sites deployed by BIH CUBI, we make the CADD, SpliceAI, MMSp, and dbscSNV scores available.
+- In priorization mode, ORPHA and DECIPHER terms are now selectable.
 
 Full Change List
 ================
@@ -66,6 +67,7 @@ Full Change List
 - Displaying warning if priorization is not enabled when entering HPO terms.
 - Added possibility to import "extra annotations" for display along with the variants.
 - On sites deployed by BIH CUBI, we make the CADD, SpliceAI, MMSp, and dbscSNV scores available.
+- In priorization mode, ORPHA and DECIPHER terms are now selectable.
 
 -------
 v0.21.0